chore: 기존 onboarding 컴포넌트 모두 삭제, 디자인 3.0 버전 온보딩 유지 & 선생님 온보딩 파일명 수정
