--- src/features/dashboard/components/onboarding/student-onboarding.tsx
@@ -1,167 +1,0 @@
-import { useEffect } from 'react';
-
-import { useStudentStudyRoomsQuery } from '@/features/study-rooms';
-import { StudyRoomsQueryKey } from '@/features/study-rooms/api';
-import { Icon } from '@/shared/components/ui/icon';
-import { PRIVATE } from '@/shared/constants';
-import { cn } from '@/shared/lib/utils';
-import { useMemberStore } from '@/store';
-import { useQueryClient } from '@tanstack/react-query';
-
-import { useOnboardingStatus } from '../../hooks/use-onboarding-status';
-import { OnboardingHeader } from './onboarding-header';
-import { OnboardingStep } from './onboarding-step';
-
-export const StudentOnboarding = () => {
-  const queryClient = useQueryClient();
-  const member = useMemberStore((s) => s.member);
-
-  // 역할에 따라 다른 쿼리 사용
-  const { data: studentRooms } = useStudentStudyRoomsQuery();
-
-  const rooms = studentRooms;
-  const firstRoomId = rooms?.[0]?.id;
-
-  // 주기적으로 쿼리를 무효화하여 실시간으로 완료 상태 업데이트
-  useEffect(() => {
-    // 페이지 포커스 시에도 즉시 체크
-    const handleFocus = () => {
-      queryClient.invalidateQueries({
-        queryKey: StudyRoomsQueryKey.studentList,
-      });
-      if (firstRoomId) {
-        queryClient.invalidateQueries({ queryKey: ['studyNotes'] });
-        queryClient.invalidateQueries({ queryKey: ['qnaList'] });
-      }
-    };
-
-    window.addEventListener('focus', handleFocus);
-    return () => {
-      window.removeEventListener('focus', handleFocus);
-    };
-  }, [queryClient, firstRoomId]);
-
-  // 학생 온보딩 상태 조회
-  const { hasRooms, hasNotes, hasQuestions, hasAssignments } =
-    useOnboardingStatus({ rooms });
-
-  // 학생의 수업노트 확인 여부
-  const hasNotesViewed = hasNotes;
-
-  const studentSteps = [
-    {
-      step: 1,
-      title: '선생님 초대 받기',
-      description: '선생님으로부터 초대 링크를 받아 스터디룸에 참여하세요',
-      icon: Icon.Mail,
-      action: '초대 링크 입력하기',
-      href: '#', // TODO: 초대 링크 입력 페이지
-      completed: hasRooms, // 스터디룸에 참여했으면 완료
-      color: 'from-gray-scale-gray-1 to-gray-scale-gray-5',
-    },
-    {
-      step: 2,
-      title: '수업노트 확인하기',
-      description: '선생님이 올려주신 수업 자료와 수업노트를 확인하세요',
-      icon: Icon.Notebook,
-      action: '수업노트 보기',
-      href: hasRooms ? PRIVATE.ROOM.DETAIL(rooms?.[0]?.id ?? 0) : '#',
-      completed: hasNotesViewed, // 학생의 경우 수업노트가 있으면 확인한 것으로 간주
-      disabled: !hasRooms,
-      color: 'from-orange-scale-orange-5 to-orange-scale-orange-10',
-    },
-    {
-      step: 3,
-      title: '과제 제출하기',
-      description: '선생님이 내준 과제를 제출하고 피드백을 받으세요',
-      icon: Icon.BookText,
-      action: '과제 확인하기',
-      href: '#', // TODO: 과제 목록 페이지
-      completed: hasAssignments,
-      disabled: !hasRooms,
-      color: 'from-orange-scale-orange-10 to-orange-scale-orange-20',
-    },
-    {
-      step: 4,
-      title: '질문하기',
-      description: '궁금한 점을 질문하고 선생님의 답변을 받으세요',
-      icon: Icon.Question,
-      action: '질문하기',
-      href: '#', // TODO: 질문 작성 페이지
-      completed: hasQuestions,
-      disabled: !hasRooms,
-      color: 'from-orange-scale-orange-20 to-orange-scale-orange-30',
-    },
-  ];
-
-  const greeting = `${member?.name || ''} 학생님, 환영합니다!`;
-  const subtitle = '선생님과 함께하는 학습 공간을 시작해보세요';
-
-  const completedCount = studentSteps.filter((step) => step.completed).length;
-  const totalSteps = studentSteps.length;
-  const nextStep = studentSteps.find((step) => !step.completed)?.step;
-
-  return (
-    <>
-      <OnboardingHeader
-        greeting={greeting}
-        subtitle={subtitle}
-        completedCount={completedCount}
-        totalSteps={totalSteps}
-      />
-      <div className="space-y-4">
-        {studentSteps.map((step, index) => {
-          const isCompleted = step.completed;
-          const isLocked = index > 0 && !studentSteps[index - 1]?.completed;
-
-          return (
-            <div
-              key={step.step}
-              className={cn(
-                'group relative rounded-2xl border-2 bg-white p-6 transition-all md:p-8',
-                isCompleted
-                  ? 'border-orange-scale-orange-30 bg-orange-scale-orange-1'
-                  : isLocked
-                    ? 'border-gray-200 bg-gray-50 opacity-60'
-                    : 'border-gray-200 hover:border-[#ff4500] hover:shadow-lg'
-              )}
-            >
-              <OnboardingStep
-                step={step.step}
-                IconComponent={step.icon}
-                status={
-                  step.step === nextStep
-                    ? 'next'
-                    : step.completed
-                      ? 'completed'
-                      : 'locked'
-                }
-                title={step.title}
-                description={step.description}
-                action={step.action}
-                href={step.href}
-                color={step.color}
-              />
-
-              {/* 단계 사이 연결선 (모바일에서만) */}
-              {index < studentSteps.length - 1 && (
-                <div className="flex justify-center pt-2 md:hidden">
-                  <div
-                    className={cn(
-                      'text-2xl',
-                      studentSteps[index + 1]?.completed || step.completed
-                        ? 'text-orange-scale-orange-30'
-                        : 'text-gray-300'
-                    )}
-                  >
-                    ↓
-                  </div>
-                </div>
-              )}
-            </div>
-          );
-        })}
-      </div>
-    </>
-  );
-};
